feat: rm unused SolutionDisplay
