+++ src/app/GameBoard/WordsGrid/utils.ts
@@ -1,0 +1,28 @@
+export const makeSolutionSets = (
+  playedWords: string[][],
+  solutionSet: string
+): Set<string>[] => {
+  const filterString = playedWords.map((word) => word.join('')).join(',');
+  const solutionsArray = solutionSet?.split('-') ?? [];
+  const filtered = solutionsArray.filter((set) => {
+    return set.includes(filterString);
+  });
+  const remainders = filtered.map((set) => {
+    const slice = set.split(',').slice(playedWords.length);
+    return slice;
+  });
+  const numToCreate = 4 - playedWords.length;
+  const sets: Set<string>[] = Array.from(
+    Array(numToCreate),
+    (_arr) => new Set()
+  );
+  remainders.forEach((wordArray) => {
+    wordArray.forEach((word, idx) => {
+      const set = sets[idx];
+      if (set) {
+        set.add(word);
+      }
+    });
+  });
+  return [...sets];
+};
